requirements.txt
altair==5.5.0
annotated-types==0.7.0
anthropic==0.101.0
anyio==4.13.0
attrs==26.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
contourpy==1.3.3
coverage==7.13.5
cryptography==46.0.6
cycler==0.12.1
distro==1.9.0
docstring_parser==0.18.0
fonttools==4.62.1
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.1
greenlet==3.0.3
groq==1.2.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.27.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.4.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
loguru==0.7.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mlxtend==0.23.1
mypy==1.10.0
mypy_extensions==1.1.0
narwhals==2.18.1
numpy==1.26.4
openai==2.30.0
packaging==24.2
pandas==2.2.2
pandas-stubs==2.2.3.250527
pillow==10.4.0
playwright==1.44.0
plotly==5.22.0
pluggy==1.6.0
proto-plus==1.27.2
protobuf==4.25.9
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.7.1
pydantic-settings==2.3.4
pydantic_core==2.18.2
pydeck==0.9.1
pyee==11.1.0
Pygments==2.19.2
pyparsing==3.3.2
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2026.1.post1
referencing==0.37.0
requests==2.32.3
rich==13.9.4
rpds-py==0.30.0
ruff==0.4.4
scikit-learn==1.5.0
scipy==1.17.1
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
streamlit==1.35.0
tenacity==8.3.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
types-pytz==2026.1.1.20260304
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
watchdog==6.0.0
xgboost==2.0.3
